fix: close RCA audit findings across core, app, GPU, services, CI, and Jenkins
- fix(app): clamp Normal mouse coordinates to X10 protocol maximum (222)
  to prevent silent u16-to-u8 truncation on wide terminals (closes #65)
- fix(core): reject out-of-range SGR 256-color indices and RGB components
  via u8::try_from instead of silent truncation (closes #66)
- fix(gpu): cast operands to u64 before multiplication in GPU buffer
  sizing to prevent usize overflow on 32-bit targets (closes #67)
- fix(services): remove dead RenderModeController::set_mode wrapper that
  silently discarded #[must_use] transition result (closes #68)
- ci: add version comments to all SHA-pinned actions across 5 workflows
  for consistent dependency auditability (closes #69)
- fix(jenkins): remove duplicate disableConcurrentBuilds in JCasC
  pipeline configuration (closes #70)
- test: add 4 regression tests for mouse clamping and SGR color
  validation (889 total, was 885)

+++ crates/features/render_gpu/src/cell_data.rs
@@ -53,7 +53,7 @@ impl GpuBackend {
             return;
         }
 
-        let buf_size = (new_capacity * std::mem::size_of::<CellInstance>()) as u64;
+        let buf_size = new_capacity as u64 * std::mem::size_of::<CellInstance>() as u64;
         let buf_usage = wgpu::BufferUsages::STORAGE
             | wgpu::BufferUsages::COPY_DST
             | wgpu::BufferUsages::COPY_SRC;
@@ -336,7 +336,7 @@ pub(super) fn prepare_and_upload_dirty_rows(
     row_byte_size: usize,
 ) {
     let flush = |backend: &GpuBackend, start: usize, end: usize| {
-        let byte_offset = (start * row_byte_size) as u64;
+        let byte_offset = start as u64 * row_byte_size as u64;
         let instance_start = start * grid_cols;
         let instance_end = end * grid_cols;
         backend.queue.write_buffer(
